Cloud_6331_secondAssignment/requirements.txt
DateTime

datetime
mysql

flask



pandas




mysql
mysql-connector-python